+++ source/extensions/omni.isaac.ros2_bridge/python/tests/__init__.py
@@ -9,3 +9,4 @@
 
 from .test_commands import *
 from .test_rclpy import *
+from .test_lidar import *