+++ source/extensions/omni.isaac.ros_bridge/docs/CHANGELOG.md
@@ -9,6 +9,8 @@ Fixed
 --------
 - odometry frame matches robot's starting frame, not the world frame. 
 - horizontal and vertical aperture use camera prim values instead of computing vertical aperture
+- lidar components publish point cloud data as PCL2 messages instead of PCL
+- lidar PCL2 messages only contain points that hit
 
 Added
 -------